Refactor color into its own module.
